Create Kosaraju_algorithm (#129)
* Create Kosaraju_algorithm

The Kosaraju algorithm is an important graph algorithm to detect strongly connected components in a graph. A directed graph is strongly connected if there is a path between all pairs of vertices. A strongly connected component (SCC) of a directed graph is a maximal strongly connected subgraph. This algorithm is often the first step in solving many graph questions.

* Create Dynamic_fibonacci.py